Refactor: move protocol enum to service mod
